Improve the docs about syncing custom files.
As reported in https://github.com/lra/mackup/issues/869, the `[configuration_files]` section of the `.mackup.cfg` is not working as advertised in the `Configuration` chapter of the docs.

The users should be pointed to the section where the syncing of custom apps (and individual files) is explained until #869 has been fixed.