Update config parser to have smarter default storage engine
Currently parser always defaults to dropbox when no config found. This is useful as it allows you to bootstrap mackup easily but requires a manual restore of .mackup.cfg for any other engine. Have updated parser to select iCloud engine if mackup config found inside iCloud, otherwise dropbox continues to be the default.