Updated the readme to explain that pip on ubuntu installs locally (#826)
The mac install instructions suggest using homebrew, which installs
mackup system wide in /usr/local. For other operating systems it
suggests pip, which also installs system-wide by default. However,
on ubuntu pip is patched to install to a user's home directory by
default. That's great, but it's not automatically in the $PATH (like
/usr/local/bin is) and this can be confusing for users who aren't used
to python.